pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Angel Torres",email = "angel_torres@epam.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types (>=0.7.0,<0.8.0)",
    "anyio (>=4.11.0,<5.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "certifi (>=2025.10.5,<2026.0.0)",
    "cffi (>=2.0.0,<3.0.0)",
    "click (>=8.3.0,<9.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "dnspython (>=2.8.0,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "fastapi (>=0.120.4,<0.121.0)",
    "fastapi-cli (>=0.0.14,<0.0.15)",
    "fastapi-cloud-cli (>=0.3.1,<0.4.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "h11 (>=0.16.0,<0.17.0)",
    "httpcore (>=1.0.9,<2.0.0)",
    "httptools (>=0.7.1,<0.8.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "idna (>=3.11,<4.0)",
    "iniconfig (>=2.3.0,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "markdown-it-py (>=4.0.0,<5.0.0)",
    "markupsafe (>=3.0.3,<4.0.0)",
    "mdurl (>=0.1.2,<0.2.0)",
    "packaging (>=25.0,<26.0)",
    "pip (>=25.3,<26.0)",
    "pluggy (>=1.6.0,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pycparser (>=2.23,<3.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "pydantic-core (>=2.41.4,<3.0.0)",
    "pygments (>=2.19.2,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "rich-toolkit (>=0.15.1,<0.16.0)",
    "rignore (>=0.7.2,<0.8.0)",
    "sentry-sdk (>=2.43.0,<3.0.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "shellingham (>=1.5.4,<2.0.0)",
    "sniffio (>=1.3.1,<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "starlette (>=0.49.1,<0.50.0)",
    "typer (>=0.20.0,<0.21.0)",
    "typing-extensions (>=4.15.0,<5.0.0)",
    "typing-inspection (>=0.4.2,<0.5.0)",
    "urllib3 (>=2.5.0,<3.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "watchfiles (>=1.1.1,<2.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "wheel (>=0.45.1,<0.46.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "boto3 (>=1.40.64,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
